only collect command name, not raw input. add option to disable collection

+++ localstack/cli/localstack.py
@@ -5,6 +5,7 @@ import sys
 from multiprocessing import Process
 from typing import Dict, List, Optional
 
+from localstack import config
 from localstack.utils.analytics.client import AnalyticsClient
 from localstack.utils.analytics.events import Event, EventMetadata
 from localstack.utils.analytics.metadata import get_session_id
@@ -46,10 +47,10 @@ def _setup_cli_debug():
     setup_logging()
 
 
-def _publish_cmd_as_analytics_event():
+def _publish_cmd_as_analytics_event(cmd: str):
     event = Event(
         name="cli_cmd",
-        payload={"raw_cmd": " ".join(sys.argv[1:])},
+        payload={"cmd": cmd},
         metadata=EventMetadata(
             session_id=get_session_id(),
             client_time=str(datetime.datetime.now()),  # TODO: consider using utcnow()
@@ -64,14 +65,17 @@ def _publish_cmd_as_analytics_event():
 @click.option("--debug", is_flag=True, help="Enable CLI debugging mode")
 @click.option("--profile", type=str, help="Set the configuration profile")
 def localstack(debug, profile):
-    publish_cmd_process = Process(target=_publish_cmd_as_analytics_event)
-    publish_cmd_process.start()
-
     if profile:
         os.environ["CONFIG_PROFILE"] = profile
     if debug:
         _setup_cli_debug()
 
+    if config.DISABLE_EVENTS:
+        return
+    ctx = click.get_current_context()
+    cmd = ctx.invoked_subcommand
+    publish_cmd_process = Process(target=_publish_cmd_as_analytics_event, args=(cmd,))
+    publish_cmd_process.start()
     publish_cmd_process.join(ANALYTICS_API_RESPONSE_TIMEOUT_SECS)
     publish_cmd_process.terminate()
